memory_manager.c: simplify logic for usedpools
I realized there's really no point in adding pools back into the
usedpools table if they have freed blocks-- we only access freed blocks
through the freeblocks table, so using "pool must have have freed blocks"
as a criterion for adding pools to the usedpools table is pointless.

Now, pools are added to the usedpools table when they are initially
carved out (as before), but they are removed only when their last unused
block is carved out. Since we can't "un-carve" a block, a pool removed
from the usedpools table will never be added back in. This means that
the only way to access blocks in a pool after this point is through the
freeblocks table, if blocks from the pool become freed.